Update install-appdynamics.sh
docker compose file fails to install app agents with "invalid login/password" with correct
credentials 
Following IF condition disabled as a temporary fix, refer reason below
###########################################
    if [ ! "$SSO_SESSIONID" ]; then
      echo "Incorrect Login/Password"
      exit
 #########################################



While sso-sessionid is not a string available under cookies.txt and thus we are failing. If you will only run wget then it would go as success. Even if we use --keep-session-cookies we would add JSESSIONID but not sso-sessionid. Workaround for now would be to remove/alter the if condition mentioned.